+++ springappuser/src/main/java/com/examly/springappuser/model/LoginDTO.java
@@ -12,10 +12,10 @@ import lombok.Setter;
 @AllArgsConstructor
 @NoArgsConstructor
 public class LoginDTO {
-    
+   
     private String token;
     private String username;
     private String userRole;
-    private int userId;
+    private Long userId;
     
 }
